Move generators and utils to their own files
