+++ zum/cli/core.py
@@ -5,8 +5,8 @@ A module to route the zum CLI traffic.
 from argparse import ArgumentParser
 from typing import Any, List, Optional
 
-import zum
-from zum.constants import DEFAULT_CONFIG_FILE_NAME
+from zum.cli.generators import generate_config_file_parser, generate_main_parser
+from zum.cli.utils import log
 from zum.engine import Engine
 
 
@@ -15,71 +15,21 @@ def dispatcher(*args: Any, **kwargs: Any) -> None:
     Main CLI method, recieves the command line action and dispatches it to
     the corresponding method.
     """
-    file_parser = generate_file_parser()
+    config_file_name = get_config_file_name(*args, **kwargs)
 
-    file_args, _ = file_parser.parse_known_args(*args, **kwargs)
+    engine = Engine(config_file_name)
 
-    engine = Engine(file_args.file)
-
-    parser = generate_parser(engine.actions)
+    parser = generate_main_parser(config_file_name, engine.actions)
     parsed_args = parser.parse_args(*args, **kwargs)
 
     engine.execute(parsed_args.action[0], parsed_args.params)  # pragma: nocover
     log(engine.output)  # pragma: nocover
 
 
-def generate_file_parser() -> ArgumentParser:
-    # Create parser
-    parser = ArgumentParser(add_help=False)
+def get_config_file_name(*args, **kwargs):
+    """Creates the config file parser and extracts the config file name."""
+    file_parser = generate_config_file_parser()
 
-    # Add file command
-    parser.add_argument(
-        "-f",
-        "--file",
-        dest="file",
-        default=DEFAULT_CONFIG_FILE_NAME,
-        help=f"config file name, defaults to '{DEFAULT_CONFIG_FILE_NAME}'",
-    )
+    file_parser_args, _ = file_parser.parse_known_args(*args, **kwargs)
 
-    return parser
-
-
-def generate_parser(actions_list: List[str]) -> ArgumentParser:
-    """Generates the action parser."""
-    warning = "Beware! No config file was found!" if not actions_list else ""
-
-    # Create parser
-    parser = ArgumentParser(
-        description="Command line interface tool for zum.", epilog=warning
-    )
-
-    # Add version command
-    parser.add_argument(
-        "-v",
-        "--version",
-        action="version",
-        version=f"zum version {zum.__version__}",
-    )
-
-    # Add file command
-    parser.add_argument(
-        "-f",
-        "--file",
-        dest="file",
-        default=DEFAULT_CONFIG_FILE_NAME,
-        help=f"config file name, defaults to '{DEFAULT_CONFIG_FILE_NAME}'",
-    )
-
-    if actions_list:
-        # Add action argument
-        parser.add_argument("action", choices=actions_list, nargs=1)
-
-        # Add params
-        parser.add_argument("params", nargs="*")
-
-    return parser
-
-
-def log(data: Optional[str]) -> None:
-    """Logs a string to the console."""
-    print(data)
+    return file_parser_args.file
